Fix Jekyll Issue
Fix issue that made the site unable to run due to a recent Jekyll change. Changed the name of "_sass/main.scss" to "_sass/base.scss" and updated the import statement in "assets/css/main.scss" accordingly. Also deleted "assets/css/variables.scss" and removed the import statement for it since it was just a blank file.